requirements.txt
torch-geometric>=2.0.0
networkx>=2.6.0
numpy>=1.19.0
scikit-learn>=0.24.0
pandas>=1.3.0
tqdm>=4.62.0
matplotlib>=3.4.0